pyproject.toml
[tool.poetry]
name = "dhlab_toolkit"
version = "0.2.0"
description = ""
authors = ["Nasjonalbiblioteket  <dhlab@nb.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.3"
requests = "^2.31.0"
pytest = "^7.4.0"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
phonetisaurus = "^0.3.0"
google-cloud-storage = "^2.16.0"
convert-pa = "^0.1.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
